gislason

gislason natural mortality relatoinship estimate M as a function of length. M=a*length^b;

# S4 method for FLQuant,numeric
gislason(length, params, a = 0.55, b = 1.44,
  c = -1.61, ...)

Arguments

length

mass at which M is to be predicted

params

FLPar with two values; i.e. a equal to M at unit mass and b a power term; defaults are a=0.3 and b=-0.288

a

0.55

b

1.44

c

-1.61

...

any other arguments

See also

lorenzen

Examples

## Not run: ------------------------------------
# params=lhPar(FLPar(linf=111))
# len=FLQuant(c( 1.90, 4.23, 7.47,11.48,16.04,20.96,26.07,31.22,
#                36.28,41.17,45.83,50.20,54.27,58.03,61.48,64.62),
#              dimnames=list(age=1:16))
# gislason(length,params)
## ---------------------------------------------