Lopt

Lopt, the length at which a cohort achives its maximum biomass, can be used as a reference point to identify growth over- or underfishing. Since taking fish below or above this size results in potential loss of yield. The total biomass of a cohort changes through time as a result of gains due to an increase in mean size-at-age and losses due to natural mortality. Lopt can therefore be estimated from the natural mortality and weight-at-age vectors.

# S4 method for FLPar
lopt(
  params,
  m = function(length, params) exp(0.55) * (length^-1.61) %*% (params["linf"]^1.44)
    %*% params["k"],
  growth = FLife::vonB,
  ...
)

Arguments

params

an FLPar object with parameter values for the natural mortality and growth functions, and the exponent b of the length/weight relationship.

m

natural mortality function, by default Gislason

growth

length or weight-at-age function, by default von Bertalanffy

...

any other arguments

Value

FLPar with $L_opt$ the length at which a cohort achives its maximum biomass

Details

Lopt is a function of growth and natural mortality-at-age and there are several approximations such as \(2/3 L_{\infty}\) and \(L_{\infty}\frac{3}{3+k/m}\). If the life history parameters and relationships are known then $L_opt$ can be found by finding the time (t) and hence length at which the maximum biomass is achieved i.e. \(L(T)^a e^{\int_0^T m(t)}\) where \(m(t)\) can be found from the relationship of mortality at length using the relationship of Gislason, assuming the von Bertalanffy growth curve.

See also

Examples

if (FALSE) {
params=lhPar(FLPar(linf=100,k=0.1,t0=-0.1,b=3))
lopt(params)
}