lorenzen

Lorenzen natural mortality relationship estimate M as a function of weight. M=a*wt^b;

# S4 method for FLQuant,FLPar
lorenzen(wt, params, ...)

Arguments

wt
mass at which M is to be predicted

params
an FLPar with two values; i.e. a equal to M at unit mass and b a power term; defaults are a=0.3 and b=-0.288

...
any other arguments

See also

Examples

if (FALSE) {
mass=FLQuant(c( 1.90, 4.23, 7.47,11.48,16.04,20.96,26.07,31.22,
               36.28,41.17,45.83,50.20,54.27,58.03,61.48,64.62),
             dimnames=list(age=1:16))
lorenzen(mass)
}