Natural mortality

Methods to provide estimates of natural mortality based on growth and reproduction parameters

roff(params, ...)

Arguments

params

FLPar

...

any other arguments

Value

returns an object of FLQuant

Details

Natural Mortality For larger species securing sufficient food to maintain a fast growth rate may entail exposure to a higher natural mortality @gislason2008does. While many small demersal species seem to be partly protected against predation by hiding, cryptic behaviour, being flat or by possessing spines have the lowest rates of natural mortality @griffiths2007natural. Hence, at a given length individuals belonging to species with a high $$L_{\infty}$$ may generally be exposed to a higher M than individuals belonging to species with a low $$L_{\infty}$$.

$$ log(M) = 0.55-1.61log(L) + 1.44log(L_{\infty}) + log(k)$$

Functional forms

Many estimators have been propose for M, based on growth and reproduction,

Age at maturity $$M=\frac{1.521}{a_{50}^{0.72}}-0.155$$ $$M=\frac{1.65}{a_{50}}$$

Growth $$M=1.5k$$ $$M=1.406W_{\infty}^{-0.096}k^{0.78}$$

$$M=1.0661L_{\infty}^{-0.1172}k^{0.5092}$$ Growth and length at maturity

$$M=3kL_{\infty}\frac{(1-\frac{L_{50}}{L_{\infty}})}{L_{50}}$$ $$M=\frac{\beta k}{e^{k(a_{50}-t_0)}-1}$$

Varing by length, weight or age

See also

gislason, lorenzen, chen

Examples

## Not run: ------------------------------------
# params=FLPar(FLPar(linf=120,k=.15,t0=-0.1,l50=60,a=0.0001,b=3))
# age=FLQuant(1:10,dimnames=list(age=1:10))
# 
# 
# roff(params)
# rikhter(params)
# rikhter2(params)
# griffiths(params)
# djababli(params)
# jensen(params)
# jensen2(params)
## ---------------------------------------------