refs

Returns a variety of reference points based on spawner per recruit (S/R) and stock recruitment relationships.

$F_MSY$ corresponds to the level of exploitation that provides the maximum, derived from a spawner and yield curves combined with a stock recruitment relationship, while $F_Crash$ is the level of F that will drive the stock to extinction.

Reference points can also be derived from the spawner and yield curves alone. For example $F_0.1$ is a proxy for $F_MSY$ and is the fishing mortality that corresponds to a point on the yield per recruit curve where the slope is 10% of that at the origin; $F_Max$ is the maximum of the yield per recruit and $SPR_O$ is the spawner per recruit at virgin biomass, and $SPR_30$ corressponds to the point on the curve where SPR is 30% of $SPR_0$. In these cases the biomass, ssb and yield values are derived by multiplying the per recruit values by the average recruitment.

refs(object, ...)

Arguments

object
FLStock

Value

data.frame

Examples

# NOT RUN {
 library(FLCore)
 library(FLBRP)
 library(ggplotFL)
 library(FLife)
 library(popbio)
 library(plyr)

 data(ple4)

 refs(ple4)
# }