Derives an FLBRP from life history parameters

Takes an FLPar object with life history and selectivity parameters and generates a corresponding FLBRP object. Can use a range of functional forms.

# S4 method for FLPar
lhEql(
  params,
  growth = FLife::vonB,
  m = "gislason",
  sr = "bevholt",
  mat = logistic,
  sel = dnormal,
  range = c(min = 0, max = 40, minfbar = 1, maxfbar = 40, plusgroup = 40),
  spwn = 0,
  fish = 0.5,
  midyear = 0.5,
  ...
)

Arguments

params

an FLPar object with life history parameters

growth

function that takes an FLPar object with parameters, by default vonB

m

character takes the natural mortality model name, by default gislason

sr

character value, "bevholt" by default

mat

function that takes an FLPar object with parameters, by default logistic

sel

function that takes an FLPar object with parameters, by default dnormal

range

numeric with age range by default from 0 to 40

spwn

numeric give propotion of year when spawning occurs, by default is params["a50"]-floor(params["a50"])

fish

numeric give propotion of year when fishing occurs, by default 0.5

midyear

when growth measured, default 0.5

...

any other arguments

units

character for vectors in FLBRP returned by method

Value

FLBRP object

See also

Examples

if (FALSE) {
data(teleost)
alb=teleost[,"Thunnus alalunga"]
eql=lhEql(lhPar(alb))
}