Derives an FLBRP from life history parameters

Takes an FLPar object with life history and selectivity parameters and generates an corresponding FLBRP object. Can uses a range of functional forms

# S4 method for FLPar
lhEql(params, growth = FLife::vonB, m = function(length,
  params) exp(0.55) * (length^-1.61) %*% (params["linf"]^1.44) %*%
  params["k"], mat = FLife::logistic, sel = FLife::dnormal,
  sr = "bevholt", range = c(min = 0, max = 40, minfbar = 1, maxfbar = 40,
  plusgroup = 40), spwn = c(params["a50"] - floor(params["a50"])),
  fish = 0.5, units = if ("units" %in% names(attributes(params)))
  attributes(params)$units else NULL, midyear = 0.5, ...)

Arguments

params
an FLPar object with life history parameters

growth
A function that takes an FLPar oject with parameters, by default vonB

m
A function that takes an FLPar oject with parameters, by default gislason

mat
A function that takes an FLPar oject with parameters, by default logistic

sel
A function that takes an FLPar oject with parameters, by default dnormal

sr
A character value, "bevholt" by default

range
A numeric with age range by default from 0 to 40

spwn
A numeric give propotion of year when spawning occurs, by default is params["a50"]-floor(params["a50"])

fish
A numeric give propotion of year when fishing occurs, by default 0.5

units
Acharacter for vectors in FLBRP returned by method

midyear
when growth measured, default 0.5

...
any other arguments

Value

FLBRP object

See also

Examples

# NOT RUN {
data(teleost)
alb=teleost[,"Thunnus alalunga"]
eql=lhEql(lhPar(alb))
# }