Generates life history parameters

Uses life history theory to derive parameters for biological relationships, i.e. or growth, maturity, natural mortality. Selectivity by default is set so age at peak selectivity is the same as age at 50/ length of the von Bertalannfy growth equation.

lhPar(params, t0 = -0.1, a = 3e-04, b = 3, ato95 = 1, sl = 2,
  sr = 5000, s = 0.9, v = 1000)

Arguments

params
FLPar object with parameters for life history equations and selection pattern. Need Linfinity to estimate other parameters, if any other parameters supplied in code then these are not provided by the algorithm

t0
of von Bertalanffy. This is a default that isnt normally derived from life history theory, as are the following args.

a
coefficient of length weight relationship

b
exponent of length weight relationship

ato95
age at which 95% of fish are mature, offset to age at which 50% are mature

sl
selectivity-at-age parameter, standard deviation of lefthand limb of double normal

sr
stock recruitment relationship

s
steepness of stock recruitment relationship

v
virgin biomass

Value

object of class FLPar with missing parameters calculated from life history theory

See also

loptAge, lhRef, lhPar, lhEql

Examples

# NOT RUN {
lhPar(FLPar(linf=200))
# }