matdd

Logistic ogive for proportion mature-at-age, modified to explicitly included maturity as a function of numbers in a cohort, i.e. density dependence

# S4 method for FLQuant,FLPar
matdd(age, params, scale, k = 1, flagAge = TRUE)

Arguments

age
ages

params
an FLPar with two values; i.e. a equal to M at unit mass and b a power term; defaults are a=0.3 and b=-0.288

scale
reference

k
rate of change in density dependence

flagAge
default is FALSE, i.e. density dependence is based on length ratherv than age

...
other arguments, such as scale, e.g. stock numbers now relative to a reference level, e.g. at virgin biomass and k steepness of relationship

Details

The relationship can be explained by population density, since as fish grow they also die and so there is potentially less competition for resources between larger and older fish. Density dependence can be modelled by a logistic function, a sigmoid curve (or S shaped) curve, with equation

f(x)=L/(1+exp(-k(x-x0)))

where e is the natural logarithm base (also known as Euler's number), x0 is the x-value of the sigmoid's midpoint, L is the curve's maximum value, and k the steepness of the curve.

Combining the two functions gives

O=aL/(1+exp(-k(n-ref)))*wt^b;

See also

Examples

# NOT RUN {
library(FLBRP)
library(FLife)

data(teleost)
par=teleost[,"Hucho hucho"]
par=lhPar(par)
hutchen=lhEql(par)

scale=stock.n(hutchen)[,25]%*%stock.wt(hutchen)
scale=(stock.n(hutchen)%*%stock.wt(hutchen)%-%scale)%/%scale

mat=matdd(ages(scale),par,scale,k=.5)

ggplot(as.data.frame(mat))+
   geom_line(aes(age,data,col=factor(year)))+
   theme(legend.position="none")+
   scale_x_continuous(limits=c(0,15))

# }